refactor TodoTest and UserTest
